EES-1550 Fix MapBlocks breaking frontend SSR rendering
